First prototype for single user mode
Undo and redo work as expected, this was the main goal of this commit.
What remains is making undo redo work with two users working at the same
time as well. This will probably imply adding something like
`CachedResource::rebuildCachesFrom(iter)` to update only what is needed
when the other user makes an undo request. The underlying container for
the history in `DrawHistory` will probably be changed for `std::list`.